backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pymysql==1.1.1
python-socketio==5.12.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.10.4
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.26.0 